installer: adjust the size of boot disk dynamically
CQID: WIND00414586

Originally we set the default size of boot disk to 100M
statically. But when the size of files that will be
copied is more than 100M, the installation can not
continue. So we will set the default size of boot disk
dynamically according to the size of copied files.

Signed-off-by: yanjun.zhu <yanjun.zhu@windriver.com>