libreport: add libproxy to DEPENDS and RDEPENDS
[WIND: US21002]

libreport checks libproxy.so dynamically, thus it would cause errors at
do_rootfs if the libproxy.so is in libpreport.rpm's depends but the
libproxy.rpm is not ready. It's not easy to disable the libproxy from
librepot or use PACKAGECONFIG, so we can easily add libproxy to the
DEPENDS and RDEPENDS to fix the problem.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>