wr-installer-rootfs.bbclass: work when installer and build dir are the same
[WIND: US21002]

This bbclass set INSTALLER_TARGET_BUILD to WRL_TOP_BUILD_DIR when we
don't specify with-installer-target-build, but it didn't work since it
would run "make bbc BBCMD='bitbake -e'" in WRL_TOP_BUILD_DIR while we
were running building there, we can't run two bitbake in the same build
dir, this patch fixes the problem.

BTW., I think that we can avoid running bitbake in the
INSTALLER_TARGET_BUILD since we have the installer-support template, we
can write out the info the template, we may do this in WRL7.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>