wr-installer-rootfs.bbclass: add the ability to add ks.cfg
[WIND: US21002]

We need a way to add the ks.cfg to the image since we support the ks
installs, now we can add it by either of:
* Set KICKSTART_FILE
* Put it in INSTALLER_TARGET_BUILD when it is a directory
* Put it in WRL_TOP_BUILD_DIR

The second or third way doesn't look very well, I'm OK to drop them.

Now if we have a ks file and add the "liveinst" to the boot args, then
it can start the ks installs.

I also added a installer-config in the rootfs, I think that we can put
the installer related info there in the future.

I think that we also need a way to easily set the boot args.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>