anaconda: check metacity before run it
[WIND: US21002]

The metacity's schemas doesn't work, thus we would get this error:

GLib-GIO-ERROR **: Settings schema 'org.gnome.metacity' is not nstalled

We did have /usr/share/glib-2.0/schemas/org.gnome.metacity.gschema.xml,
we can run "glib-compile-schemas" to fix it, but the problem is that we
we fix it, progress bar will became ugly, so we will remove metacity
from anaconda's RDEPENDS and check whether /usr/bin/metacity before run
it to workaround the problem atm.

We still keep the metacity recipe in case we can fix the look.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>