system-config-keyboard: add gettext-native as a DEPENDS
CQID: WIND00439864

system-config-keyboard requires gettext to build yet it currently doesn't list
gettext as a direct dependency. Build failures have been observed on occassion
because of this so adding gettext as a DEPENDS to take the chance out of
building things.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>

Tweak to append 'gettext-native' instead of 'gettext' to the DEPENDS.

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>