wr-installer-rootfs.bbclass: remove wrl_installer_check and wrl_installer_stamp
[Issue: LIN6-7253]

The wrl_installer(), wrl_installer_check() and wrl_installer_stamp() has
a few duplicated lines and cause confusions, I think that these are
because of the previous initramfs, the initramfs is only used for
booting and switching the rootfs now, we can remove the duplications and
only keep wrl_installer(), and 61 lines has been removed.

Add wrl_installer_copy_pkgs() which is used for copying the rpm packages
to the rootfs when INSTALLER_TARGET_BUILD is a directory.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>