anaconda: run sync() before shutdown or reboot
[Issue: LIN6-7301]

We need flush file system buffers before shutdown or reboot, otherwise
the data may incomplete.

I'm a little curious that we need run this, but I did meet the error one
or two times that the installer told everything is OK, but the
bootloader can't start, the testing team also met the same problem by
accidentally, maybe because of some kernel configurations.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>