anaconda: EFIGRUB should base on GRUB2
[Issue: LIN6-7331]

We use grub2 and grub2-efi by default, so the EFIGRUB class should base
on GRUB2, otherwise it doesn't work. Now we only have EFIGRUB which is
used for grub2 in fact, we may need both EFRGRUB2 and EFIGRUB if we make
sure that we can support the grub (and the efi) well.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>