smartinstall.py: run depmod as image.bbclass
[Issue: LIN6-7462]

The oe-core doesn't run the depmod for each kernel-module during
do_rootfs, but run it once in image.bbclass, we need do the similar
thing in smartinstall.py.

Note, the /etc/init.d/modutils.sh can run the depmod, but it may be a
little ater, for example it runs after udev which causes we get no
modules at firstboot.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>