anaconda: change /bin/sh to relative symlink
Issue: LIN6-7510

The anaconda requires /bin/sh be a relative symlink, but we use absolute
symlink, for example:
       /bin/sh -> /bin/bash

Change it to /bin/sh -> bash when possible, and show error messages when
can't change.

We'd better fix it from bash or busybox, but it's not easy to fix that
in WRL6, the update-alternatives will change it to absolute symlink even
if we use the relative one.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>