tigervnc: compilation break without cast
The code is currently correct C99, which is what the Xorg code is written in.
This patch is only needed for forks like TigerVNC which use this code in C++
programs.  I don't mind accepting small patches like this, especially since I
made the exact same patch to the Solaris packages so we can build TigerVNC,
but we're not going to guarantee our C code will always be portable to C++.

Reviewed-by: Alan Coopersmith <alan.coopersmith at oracle.com>

http://lists.x.org/archives/xorg-devel/2015-January/045096.html
Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>