python-smartpm: limit status updating frequency
While installation with text mode, the status refresh is
too frequency which caused the main process was blocked
and wasted times.

The python-smartpm has supported to control the frequency,
but it did not work for wr-installer.

The fix makes it work, although it lost some status, but
it is ok for us.

The refresh frequency is time/1s

Issues: US55695

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>