lvm2.bbappend: fix QA warning and disable lvm2.service for systemd
We add script build_lvm_initramfs.sh in lvm2 for both of target
image and host image. In this place, we tweak bbapend to add it
to FILES_${PN} for host image, which avoids not shipped warning.

The host image do not need to run 'vgchange -ay' at boot time,
which could conflict with kickstart anaconda, that the racing
about existed logic volume.

Rally: US55699

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>