anaconda: fix lvm boot failed while package group selection
While partition rootfs as lvm, we need to invoke build_lvm_initramfs.sh
to generate initramfs. It requires the packages ldd, gzip and iputils to
be installed to target image.

So while customize to select packages, we need to have them to make
sure the target could boot even though the user does not select them.

Rally: US55698

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>