anaconda: fix exception while invoking _anyRepoEnabled
Invoking _anyRepoEnabled to test whether the enabled repository is available,
while no repo is available(such as do not add any network repo in PXE), an
exception from _anyRepoEnabled break anaconda.
...
|Traceback (most recent call first):
|  File "/usr/lib64/python2.7/site-packages/pyanaconda/iw/task_gui.py",
line 600, in _anyRepoEnabled
|    if model.get_value(iter, 0):
|  File "/usr/lib64/python2.7/site-packages/pyanaconda/iw/task_gui.py",
line 524, in getNext
|    if not self._anyRepoEnabled():
|  File "/usr/lib64/python2.7/site-packages/pyanaconda/gui.py", line 1208,
in nextClicked
|    rc = self.currentWindow.getNext ()
|TypeError: iter must be a GtkTreeIter
...

We should use False to instead of exception in this situation.

Issues: LINBDF7-13

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>