anaconda: list all ipaddr while start vnc
While doing text mode installation and select to start vnc,
the ipaddr which serial port shows is '127.0.0.1'.
...
|08:16:34 Please manually connect your vnc client to 127.0.0.1:1
to begin the install.
|08:16:34 Starting graphical installation.
...

It is not helpful for the user to start vncview to connect vnc server.
We list all ipaddr while start vnc. Here is the output:
...
|08:16:34 Please manually connect your vnc client to 127.0.0.1:1,
128.224.162.181:1 to begin the install.
|08:16:34 Starting graphical installation.
...

Issues: LINBDF7-34

The net implements refers:
http://code.activestate.com/recipes/439093-get-names-of-all-up-network-interfaces-linux-only

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>