wr-x86-server: fix CDROM is not detected issue
Rally: TA95170

On HP Proliant servers, installer ISO is usually connected via virtual
CDROM with iLO. But we found that /run/media/sr0 is not detected when
system try to mount real rootfs in initramfs. The root cause is when
trying to mount real rootfs, /dev/sr0 is not available yet, so adding
some sleeping time to make sure /dev/sr0 is available.

Signed-off-by: Yunguo Wei <yunguo.wei@windriver.com>