systemd: Adding local udev rules to avoid mouse issue on guest installer
We found ISO as the guest install media, mouse doesn't work on anaconda.
That is because ovirt will add "-device usb-tablet" to qemu options.
However, this device is not needed in our system. But this device will
cause evdev moudule in xorg driver unloaded, so as it runs into some
power managerment issue.

The original hid pm rule for qemu usb tablet is introduced by commit
d74a0ecdbc85a482cab6e7eae8dcb48185d44d84 in oe-core. But that rule
doesn't work with HP certification prject on usb tablet, so we skip it
for now.

This is not a fix but a workround. We need to figure out the root cause
in future.

Signed-off-by: Yunguo Wei <yunguo.wei@windriver.com>