anaconda-init: Adding barrier to make sure nm is up before anaconda-init starts
We use the local anaconda-init to add barrier, otherwise the network
drivers' messages in console will mess up the prompt line when it's
asking you what to install.

This is backport from 6.0 installer, but network manager is depreciated
in 7.0, so adding sleepting time to make sure all networking driver's
message is printed in advance.

Signed-off-by: Yunguo Wei <yunguo.wei@windriver.com>