iscsi-initiator-utils: fix QA issues
Issue: LIN8-1864

Fixed:
WARNING: QA Issue: iscsi-initiator-utils:
/iscsi-initiator-utils-tests/opt/iscsi-initiator-utils-tests/isns-tests/data/test08/01-pauw1
is owned by uid 1786, which is the same as the user running bitbake.
This may be due to host contamination [host-user-contaminated]
WARNING: QA Issue:
/opt/iscsi-initiator-utils-tests/regression.sh_iscsi-initiator-utils-tests
contained in package iscsi-initiator-utils-tests requires /bin/bash, but
no providers found in its RDEPENDS [file-rdeps]

It uses 'cp -a' to install the files, so fix the owner to root:root
and add RDEPENDS on bash for the -tests package.

Signed-off-by: Jackie Huang <jackie.huang@windriver.com>