support ISO mediacheck for anaconda installer
Mediacheck is a tool we use to test the integrity of ISO images.  The
ISO9660 specification allows for a 512 byte region which is for
application use.  We use it to store a checksum of the ISO image.  Of
couse putting the checksum into this region will change the checksum
of the final ISO image, so when we verify the checksum later we have
to fill in this 512 region with the original contents before the
implantation. It is set to all ASCII 32 decimal spaces.

If you have a set of ISO images you can implant the checksum data with
this command:

  implantmd5iso <isoname>

NOTE: You cannot re-implant an ISO once its been implanted.

To check a ISO image you can use this command line:

   checkisomd5 <isoname>

The tools are in the isomd5sum package.

Issue: US65409

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>