python-smartpm: fix smart update failed while http proxy used
While using http proxy, it missed to add pycurl.HTTPPROXYTUNNEL:
...
When  an  HTTP proxy is used (-x, --proxy), this option will cause non-HTTP
protocols to attempt to tunnel through the proxy instead of merely using it
to do HTTP-like operations. The tunnel approach is made with the HTTP proxy
CONNECT  request  and requires that the proxy allows direct connect to the
remote port number curl wants to tunnel through to.
...

Issue: US64354

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>