anaconda: detect existence of Xorg
Before anaconda starting Xorg, detect its existence.

It is helpful for debugging anaconda: only kill -9 anaconda
is enough, and do not need to kill -9 Xorg any more.

(The reason to add the detection is we have a display bug that
need to keep Xorg alive and restart anaconda by kill -9)

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>