anaconda: fix reboot failed after install finished while live install
The upstream use a separate way to reboot for live install,
because they need eject live CD after installation finished.
(They add a simple wrapper around anaconda to kick off an install
from a live CD)

But wrlinux do not need it, and could not eject installation live CD
neither. We mount rootfs to live CD (/dev/sr0), and it is hard to umount
rootfs before eject.

And we could not use upstream's wrapper, it is gnome desktop related
and not proper for wrlinux.

Issue: US81339

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>