networkmanager bbappend: remove
We do not need bbappend any more.

The networkmanager in meta-oe has been upgraded to 1.0 since 2015
...
commit dd9d09945da519fa67a5da61b48c4ddd534f078a
Author: Andreas Mï¿½ller <schnitzeltony@googlemail.com>
Date:   Mon Jun 15 11:50:55 2015 +0200

    networkmanager: update to 1.0.2

    * at_console support is gone
    * add PACKAGECONFIG for bluez5/modemmanager/ppp and disable them by default

    Signed-off-by: Andreas Mï¿½ller <schnitzeltony@googlemail.com>
...

And current version is 1.0.10, so the 0.9.8.0's bbappend has not used in a long time.
...
WARNING: No recipes available for:
  layers/wr-installer/recipes-connectivity/networkmanager/networkmanager_0.9.8.0.bbappend
...

Currently anaconda works well with networkmanager 1.0.*, and we may redefine its
bbappend in US81332 Network setting with graphic mode.

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>