classes/installer-banner: add banner message and not enable installer-support template by default
A template.conf should not be injecting another template.

Instead the user should be required to enable the template themselves.  We can
have the message printed to tell them to do this:
----------------------------------------------------------------------
WARNING: The wr-installer layer has been included, and the current
build is a regular target build. It requires to enable
feature/installer-support template. Add option
--templates=feature/installer-support to setup.sh, or set WRTEMPLATE
+= 'feature/installer-support' in local.conf
----------------------------------------------------------------------

And not print the message if they have:
----------------------------------------------------------------------
The wr-installer layer has been included, and the current build is a
regular target build. After it is finished, create a new installer
build with DISTRO = 'wrlinux-installer', INSTALLER_TARGET_BUILD =
'<target-build-topdir>|<target-build-image>' and
INSTALLER_TARGET_IMAGE = '<target-image-pn>' in local.conf
----------------------------------------------------------------------

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>