drop wr-installer-grub-efi.bbclass
Previously, use this bbclass to fix efi boot menu not work
(choose to boot from live USB, but it actually boots from
hard disk, or reverse)

We fix the issue at anaconda runtime to invoke grub-install
to generate efi filesystem with specifying the partition.

So do not need the bbclass.

Issue: LIN10-1734

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>