README: introduce how to use qemu to boot installed target image
The IDE driver in the kernel is fragile and in 4.12 is causing
backtraces. To unblock 4.12 kernel merging use the virtio CD driver
instead to mount iso images which should be faster and more stable.

Update README to introduce how to use qemu to boot installed target
image with virtio.

Issue: LIN10-2084

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>